# accounting_server/openapi.yaml
openapi: 3.1.0
info:
  title: Unified Global Ledger API
  description: |
    A centralized financial orchestration server providing an abstraction layer
    between disparate global payment ecosystems. Uses internal "Accounting Credits"
    as the unit of account for P2P transfers and value accumulation.
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@example.com
  license:
    name: Proprietary
servers:
  - url: https://api.ledger.example.com/v1
    description: Production
  - url: https://api.staging.ledger.example.com/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Local development
...
security:
  - BearerAuth: []
paths:
...
  /transfers:
    post:
      tags: [Transfers]
      summary: Initiate P2P transfer
      description: |
        Initiate a transfer between two internal accounts. This creates a pending
        transfer that must be confirmed by the recipient using the confirmation token.

        Flow:
        1. Sender creates transfer (funds are held from available balance)
        2. Sender receives confirmation_token in response
        3. Sender passes token to recipient (out-of-band)
        4. Recipient confirms transfer via POST /transfers/confirm

        Pending transfers can be cancelled by the sender via POST /transfers/{id}/cancel.
      operationId: createTransfer
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransferRequest'
      responses:
        '202':
          description: Transfer created and pending confirmation
          headers:
            Location:
              schema:
                type: string
              description: URL of the created transfer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
              example:
                id: "550e8400-e29b-41d4-a716-446655440000"
                type: transfer
                status: pending
                source_account_id: "550e8400-e29b-41d4-a716-446655440001"
                destination_account_id: "550e8400-e29b-41d4-a716-446655440002"
                amount:
                  amount: "10000"
                  currency: "CREDIT"
                confirmation_token: "txn_confirm_abc123xyz456"
                created_at: "2026-02-06T10:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: Validation error or insufficient funds
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                insufficient_funds:
                  $ref: '#/components/examples/InsufficientFunds'
                same_account:
                  $ref: '#/components/examples/SameAccount'
        '429':
          $ref: '#/components/responses/RateLimited'
  /transfers/{transferId}:
    get:
      tags: [Transfers]
      summary: Get transfer details
      operationId: getTransfer
      parameters:
        - name: transferId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Transfer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /transfers/confirm:
    post:
      tags: [Transfers]
      summary: Confirm a pending transfer
      description: |
        Confirm a pending transfer using the confirmation token. This endpoint must
        be called by the recipient (destination account owner) to complete the transfer.

        The confirmation token is provided to the recipient out-of-band by the sender.
        Upon successful confirmation, held funds are transferred to the recipient's account.
      operationId: confirmTransfer
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmTransferRequest'
      responses:
        '200':
          description: Transfer confirmed and completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Caller is not the recipient of this transfer
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Transfer not found or token invalid
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Transfer is not in pending status (already confirmed, cancelled, or expired)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '429':
          $ref: '#/components/responses/RateLimited'
  /transfers/{transferId}/cancel:
    post:
      tags: [Transfers]
      summary: Cancel a pending transfer
      description: |
        Cancel a pending transfer. Only the sender (source account owner) can cancel
        a transfer, and only while it is in 'pending' status.

        Upon cancellation, held funds are released back to the sender's available balance.
      operationId: cancelTransfer
      parameters:
        - name: transferId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Transfer cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Caller is not the sender of this transfer
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Transfer cannot be cancelled (already confirmed, cancelled, or expired)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  # ============================================
  # DEPOSITS
  # ============================================
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Authentication via Bearer token. Two formats are accepted:
        - **JWT token**: Standard JWT from authentication service (full access)
        - **API token**: Long-lived token starting with `at_` prefix (scoped access)

        API tokens can be created via POST /v1/api-tokens using JWT authentication.

  # ============================================
  # PARAMETERS
  # ============================================
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Unique key to ensure idempotent request processing (UUID v4 recommended)
      schema:
        type: string
        format: uuid
        example: "550e8400-e29b-41d4-a716-446655440000"
...
  schemas:
...
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: string
          pattern: '^\d+$'
          description: Amount in smallest currency unit (e.g., cents)
          example: "10000"
        currency:
          type: string
          description: Currency code
          example: "CREDIT"
...
    TransferStatus:
      type: string
      enum: [pending, completed, cancelled, expired]
      description: |
        Status of a P2P transfer:
        - pending: Awaiting recipient confirmation (funds held)
        - completed: Confirmed by recipient (funds transferred)
        - cancelled: Cancelled by sender (funds released)
        - expired: Confirmation token expired (funds released)
...
    Transfer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [transfer]
        status:
          $ref: '#/components/schemas/TransferStatus'
        source_account_id:
          type: string
          format: uuid
        destination_account_id:
          type: string
          format: uuid
        amount:
          $ref: '#/components/schemas/Money'
        fee:
          $ref: '#/components/schemas/Money'
        description:
          type: string
        confirmation_token:
          type: string
          nullable: true
          description: |
            Token required for recipient to confirm the transfer.
            Only present when status is 'pending'. Sender must pass this
            token to the recipient out-of-band.
          example: "txn_confirm_abc123xyz456"
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when transfer was confirmed/completed
        cancelled_at:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when transfer was cancelled
        expires_at:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when pending transfer expires if not confirmed
        metadata:
          type: object
          additionalProperties: true
    CreateTransferRequest:
      type: object
      required: [source_account_id, destination_account_id, amount]
      properties:
        source_account_id:
          type: string
          format: uuid
        destination_account_id:
          type: string
          format: uuid
        amount:
          $ref: '#/components/schemas/Money'
        description:
          type: string
          maxLength: 500
        metadata:
          type: object
          additionalProperties: true
    ConfirmTransferRequest:
      type: object
      required: [confirmation_token]
      properties:
        confirmation_token:
          type: string
          description: |
            The confirmation token received from the sender.
            This token validates that the recipient is authorized to complete the transfer.
          example: "txn_confirm_abc123xyz456"

    # --- Deposit ---
...
    ProblemDetails:
      type: object
      description: RFC 9457 Problem Details
      properties:
        type:
          type: string
          format: uri
          description: URI reference identifying the problem type
        title:
          type: string
          description: Short human-readable summary
        status:
          type: integer
          description: HTTP status code
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence
        instance:
          type: string
          description: URI reference identifying the specific occurrence
        trace_id:
          type: string
          description: Request trace ID for debugging
      additionalProperties: true

    # --- Health ---
...
  responses:
    BadRequest:
      description: Malformed request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Unauthorized:
      description: Authentication required
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Forbidden:
      description: Access denied
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'

  # ============================================
  # EXAMPLES
  # ============================================
  examples:
    InsufficientFunds:
      value:
        type: "https://api.ledger.example.com/problems/insufficient-funds"
        title: "Insufficient Funds"
        status: 422
        detail: "Account acc_abc123 has insufficient available balance. Required: 1000 CREDIT, Available: 500 CREDIT."
        instance: "/v1/transfers"
        account_id: "acc_abc123"
        required_amount: "1000"
        available_amount: "500"
    SameAccount:
      value:
        type: "https://api.ledger.example.com/problems/validation-error"
        title: "Validation Error"
        status: 422
        detail: "Cannot transfer to the same account."
        instance: "/v1/transfers"